+++ RELEASES.md
@@ -23,6 +23,7 @@ Upcoming
 - Add Akamai Sender module
 - Support multiple configuration directories for DirectoryClassLoader
 - Add rootDir property to AkamaiSender
+- Add Liquibase migration support